airflow-requirements.txt
apache-airflow-providers-amazon>=8.19
boto3>=1.34
pandas>=2.2
pyarrow>=15.0
psycopg2-binary>=2.9
pydantic>=2.6
pydantic-settings>=2.2
